+++ client/src/constants.ts
@@ -6,3 +6,7 @@ export const CRUD_MODE = JSON.parse(
   process.env.REACT_APP_CRUD_MODE ||
     JSON.stringify(process.env.NODE_ENV === "development")
 );
+
+export const BASIC_SEARCH_WIDGET = JSON.parse(
+  process.env.REACT_APP_BASIC_SEARCH_WIDGET || JSON.stringify(!CRUD_MODE)
+);
